+++ app/src/main/java/com/thequietz/travelog/place/viewmodel/PlaceSearchViewModel.kt
@@ -3,7 +3,7 @@ package com.thequietz.travelog.place.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import com.thequietz.travelog.BaseViewModel
+import com.thequietz.travelog.common.BaseViewModel
 import com.thequietz.travelog.place.model.PlaceSearchModel
 import com.thequietz.travelog.place.repository.PlaceSearchRepository
 import com.thequietz.travelog.schedule.model.SchedulePlaceModel
